added

+++ TBRPG-Project-Alpha/Assets/_Scripts/EnemyAIAction.cs
@@ -6,5 +6,5 @@ public class EnemyAIAction
 {
 
     public GridPosition gridPosition;
-    public int actionValue;
+    public float actionValue;
 }
